Prepare to deprecate unix only direct mmap code support
Adding a warning for now that will be seen if you for some
reason are not using the portable File::Map based implementation.
The current duplication is confusing and the direct mmap support
has not been actively maintained.  This is a final nudge in case
anyone still uses the legacy/non-portable implementation.