Fix sf.net bug #331 "uniq does not always return proper object"
The problem was that for some combinations of input, the output
was generated by calling pdl() directly which uses class PDL by
default.  I changed the call from pdl([$stuff]) to
PDL::pdl(ref($stuff),[$stuff]) and now the appropriate class
is obtained.  We probably should scrub the PDL sources for any
other raw uses of the pdl() constructor that should be made
object type safe.