Fix sf.net bug #338 PDL::FFT uses backwards sign convention from FFTW
Also clarified some of the docs, added some signatures for the
routines, and a WARNING at the top regarding the change in sign
convention between forward and reverse FFTs.

NOTE: The inverse FFT is normalized so that FFT followed by IFFT
are true inverses.  This is different from the unnormalized transform
calculated by FFTW.