Fix inconsistent internal state due to exceptions from out-of-bounds slices.
Out-of-bounds slices are not caught at creation time, but later during the
evaluation of RedoDims. During the RedoDims code, exceptions were thrown
after the slice piddle's internal state was partially setup, and usually
after it was setup enough to masquerade as a functional piddle. This led
to erroneous dimension size reporting, and potentially to segmentation
faults. The best solution is probably to magically swap out the slice
piddle with a "barf" slice, which would just always barf with the given
error message when any of its methods were invoked. As a stop gap, I
here reset one of the internal flags so that the RedoDims code is invoked
on any other operations on the piddle, ensuring that any data access or
dimension calculations retrip the barf.