Fix OpenBSD pthread build problem for non-threaded perls
There is an open Feature Request ticket to improve the
pthreads detection and build for PDL.  For OpenBSD 5.3
and non-threaded perls, the current build fails.  For
that case, I now disable the pthreads building.  This
can be re-enabled when the pthreads detection is revisited,
perhaps during the PDL-2.008 clean-up release.