Obfuscate pod constructs so perldoc doesn't see them
Some of the PP commands generate POD and the format of
the text pieces used to produce the generated POD looked
like valid POD since they were in heredocs at the first char.
Added a prefix and map command to hide them.  Now only PP.pod
shows as having docs.  Still need to add POD for the various
pp_xxx routines in a form so pdldoc (especially with -a will
find the references).