pdl_trans is now properly defined as a flexible array
pdl_trans was intended to have a variable number of pdl objects in it, but the
code defined this number as 1. This causes issues with gcc4.8, which is more
aggressive at optimizing the code based on what the programmer tells it. In this
case, the pdl[1] that the code had was taken literally. The new flexible array
implementation tells the compiler exactly what is happening, and all is well.

See this thread:

http://mailman.jach.hawaii.edu/pipermail/perldl/2013-October/008191.html