Fix sf.net bug #352 reorder() not an lvalue sub
This adds reorder to the list of lvalue subs in PDL::Lvalue.

NOTE: The docs for Lvalue subs in perlsub say that in an
lvalue sub you must not use return but simply have a value
that remains at the end of subroutine scope.  A number of
PDL::Slices routines actually use return $answer_pdl; and
seem to be ok.  We should probably clear this up and make
sure nothing is waiting to bite us later.