Fix modulo operation for longlong operands.
Modulo is a weird operator, and there are actually two different
macros: one for Byte and Unsigned, and one for all other datatypes.
The code that decides between the two was missing the "Q"
identifier for longlong datatypes, so it was returning garbage.

Added one test to check for this.