Fix for string parsing and dimension handling (fixes SF#367)
There was a logic error where ->setdims([]) would get called when trying
to remove the outermost dimension, but this would run after using
clump() on a string that had a single value and ->dims() == 1.

This also fixes the BAD value parsing bug in SF#367 which led to an
incorrect BAD value when parsing the string "[BAD]".
See <http://sourceforge.net/p/pdl/bugs/367/>,
<https://github.com/PDLPorters/pdl/issues/47>.