Remove -Wunsequenced clang warnings.
All of these came from one line in Basic/Gen/PP.pm that used an
assignment in a macro argument. But in the Perl macro cascade
there is #defined HvAUX(hv), which uses its argument twice.
This caused the assignment to happen twice. It was a harmless
unintended consequence, but at least the warnings have gone away.