Suppress -Wliteral-conversion warning on clang for ufunc.pd.
clang was complaining because the generated code turned out to
be somthing like 'PDL_short foo = 0.25;' which is a type mismatch.
0.25 got rounded to 0 before assignment to PDL_short foo.
This is actually what the cheesy test for floating-pointiness
intended, so I suppressed that warning in the Makefile.PL.

An alternative would be to have a type() look for integers and
one for floating-point datatypes.