add PDL_INVALID to the pdl_datatypes enum list.
The list is generated by Types.pm.PL, but gets put into pdl.h and
pdlsimple.h. This was motivated by code in Bad.c that assigns to
a variable of type pdl_datatypes the value -1 (in some error condition
that is unlikely to be reached). This jives with the ANYVAL_FROM_CTYPE
macro in pdl.h.  Thanks to Chris Marshall and David Mertens for the
pointers.  See pdl-devel 2016-05-05 discussion.