Prevent segfault when qsort etc called with a scalar argument
If a user mistakenly passed a perl scalar argument to qsort and
its ilk, the SV would be mistaken for a piddle and passed to
the underlying generic_qsort etc routines. Not surprisingly,
those routine's array index tricks would cause a segfault. This
commit causes qsort, qsorti, qsortvec, and qsortveci to barf
if an argument is passed with incorrect dimensions.

Some care is taken to allow trivial sorts.

2 tests for each routine have been added, and some existing
tests labeled.