Make code generation for multiple types happen in complexity order.
Related to SF.net patch #84 from Debian (reproducibile-build.patch)
but instead of sorting the keys of PDL datatypes lexigraphically,
using PDL::Types::typesrtkeys() sorts them in order of complexity.
Satisfies the goals of reproducible builds but does it in a way
that makes more sense for PDL in this case.