Fix typo in gsl_sf_coupling.pd
As reported on Pdl-general, there was a typo in the pp_def code
for gsl_sf_coupling_6j in which it was mistakenly calling the GSL
routine gsl_sf_coupling_3j_e instead of _6j_e.

Thanks to Yury Pakhomov for reporting the bug, and to
Chris Marshall for quickly pointing out the solution.