Try out a fix for memory freeing issue in warp2d.
There seems to be a problem with threads and Windows and freeing
memory.  Previously the warp2d test would crash on Strawberry Perl
with a "free to wrong pool" error.

I followed the advice from BrowserUK in this PerlMonks thread:
http://www.perlmonks.org/?node_id=742205 to create a pass-through
routine to free the interolation kernel's memory.  Thanks for asking
the similar question 9 years ago, Rob!

These changes don't cause any tests to fail on my Mac, we shall see
on the AppVeyor.