Fix doc db to accomodate multiple functions of the same name.
Previously the unique key for the pdldoc.db entries was the function
name. But different modules can provide the same function name;
only the last-scanned one would show up in the db.  e.g. 'max' in
PDL::Ufunc and PDL::GSL::RNG.  This wasn't a perl namespace collision,
just a shortcoming in the doc database design.

This makes the unique key the fully-qualified function name, e.g.,
'PDL::Ufunc::max'. 'apropos' now has a new column which tells the
package name.  I've introduced a few shorthands. The rest of the
changes are basically decoupling the package and the function.

Probably some of the regexs could be cleaned up, and the print
formatting too.  And the documentation needs to be updated to
reflect these changes.  But the code should work as-is.