Do not reduce stack extension in PP code by the number of inputs.
At least as far back as May of 2005, ExtUtils::ParseXS has
reduced the size of the stack pointer SP by the number of items
(i.e., input variables) at the beginning of each xsub.  XS code
generated by PDL::PP was also reducing the size of SP by the
number of items (or, at least, was extending the size of the stack
by <items> too few elements).  This commit removes the duplication.