Improve compatibility with Perl's experimental 'bitwise' feature.
Patch applied as-is, without "simply wrapping and2, or2 and xor
in an anonymous sub for overloading, as is already done with ~"

From https://rt.cpan.org/Public/Bug/Display.html?id=123901:
Perl 5.22 introduced a new ‘bitwise’ experimental feature that
splits the two behaviours of bitwise into two separate sets of
ops, effectively giving Perl three sets of bitwise ops:

  bitwise-feature: &. |. etc. (string)
  bitwise-feature: & | etc. (number)
  old bitwise ops: & | etc. (string or number)

To distinguish between the last two sets of ops, overload methods
get a fifth true argument when the ‘bitwise’ feature is enabled.
PDLs methods croak on this ‘wrong’ number of arguments.  The attached
patch fixes it, but after writing this patch I wondered whether it
wouldn’t be better simply to wrap and2, or2 and xor in an anonymous sub
for overloading, as is already done with ~.  In any case, I leave it to
you to do with this patch as you see fit. Please note that there were
no tests at all for ^ and ~ that I could find, so I added one for each.
__END_OF_PATCH_COMMENT__