Change doc db structure to three-layer format.
This revamps the previous two commits 21b101a and 5b9846e.
Having the db primary key be ModuleName::funcname was a bad idea
because it muddled the case in which the fully-qualified name was
required to call a function.  Instead, this change adds another
layer to the documentation database and hashref structure:
$hash->{funcname}->{modulename} is needed to get the function's
information.  This is much more robust, though it adds some
complexity to the interface.

I updated the documentation and the example in Doc.pm to match
the new structure.

I also updated Perldl.pm and scantree.pl for the new structure.
usage, sig, and badinfo should work just fine.

apropos still gives the (shortened) module name along with the
function name and reference text.

help more correctly handles the case of many matches (try 'help new'!).