native complex types
some tests for complex numbers added, notably in ops_complex.t, which
still has a lot of redundant tests from ops.t. I've noticed implicit
type conversion if an operation is not defined for a given type,
e.g. complex < or >; also asin(3) returns a complex number now. I'm not
sure this behaviour is desired, nor have I found out the exact piece
of code that does that. One solution would be PMCode wrappers for many
operators/functions that check types and throw errors, or not, maybe
switched on a config variable?