Squash whitespace-only string in Makefile.PL to quiet EU::MM (#276)
If both of these variables are empty, then the string here would
be just " ", and that would generate a few 'uninitialized' warnings
during 'perl Makefile.PL'. No major problem, but it looks bad.
This sets it to the empty string in that case.  Can't use "s/^ $//r" because
that's not available in Perl <5.14.