rfits properly treats BLANK in rice-compressed images with no BSCALE
As identified in issue #288, if a Rice-compressed image had a BLANK
keyword in the FITS header, that would not properly get set to BAD
if the FITS header did not also contain the BSCALE keyword.  This
was because treat_bscale(), which also deals with the BLANKs, was
not called if there was no BSCALE present.

Ideally this change would have a corresponding test, but without
compression in wfits, I don't see how to do it simply.