GHA: Restore cygwin job to build status pre-reqs
This continues to run the ci and ci-ubuntu-containers jobs only on push
and pull_request events and skips them on the tag creation event so as
to save time.

This means that the cygwin job will run on the tag creation event as
before, but will now also generate a build-status notification. The
skipped ci/ci-ubuntu-containers jobs will not interfere with the
build-status.