fix incorrect (but working under special case) context for repeat operator
A parenthesized string before the repeat operator normally indicates
a list element repitition, *except* in scalar context, where Perl
will turn it into a string concatenation repetition.  While the
existing code

        t/' . ('*/') x $_ . '*.t'

works because of this special case-ing by Perl, it is confusing to the
reader (and even the author!)

This commit changes this to

        t/' . ('*/' x $_ ). '*.t'

which explicitly uses the string concat repeat context and retains
the parenthesis so the reader doesn't have to worry about operator
precedence.